move the platform-specific parts, and interface.dart inside lib/src
