+++ lib/src/client_native.dart
