+++ lib/src/client_stub.dart
