+++ lib/src/interface.dart
